requirements.txt
aiohttp~=3.8.1
pydantic~=1.9.0
aioredis~=2.0.1
PyYAML~=6.0